+++ acc/components/samples/isotropic_box.py
@@ -6,32 +6,16 @@
 import numpy as np
 from numba import cuda, void, int64
 from numba.cuda.random import xoroshiro128p_uniform_float32, xoroshiro128p_type
-from math import sqrt, exp, pi, sin, cos
+from math import sqrt, exp
 
 from .SampleBase import SampleBase
 from ...neutron import absorb, prop_dt_inplace
 from ...geometry.onbox import cu_device_intersect_box
+from ...kernels.isotropic import scatter
 
 from ...config import get_numba_floattype
 NB_FLOAT = get_numba_floattype()
 
-
-@cuda.jit(device=True)
-def scatter(threadindex, rng_states, neutron):
-    # randomly pick direction
-    cos_t = xoroshiro128p_uniform_float32(rng_states, threadindex)*2-1
-    phi = xoroshiro128p_uniform_float32(rng_states, threadindex)*(2*pi)
-    if cos_t>1: cos_t = 1
-    sin_t = sqrt(1-cos_t*cos_t)
-    sin_p, cos_p = sin(phi), cos(phi)
-    # compute velocity
-    vx, vy, vz = neutron[3:6]
-    vi = sqrt(vx*vx+vy*vy+vz*vz)
-    vx = vi*sin_t*cos_p
-    vy = vi*sin_t*sin_p
-    vz = vi*cos_t
-    neutron[3:6] = vx,vy,vz
-    return
 
 category = 'samples'
 class IsotropicBox(SampleBase):
