+++ setup.py
@@ -16,6 +16,7 @@ setup(
         'mcvine.acc',
         'mcvine.acc.test',
         'mcvine.acc.geometry',
+        'mcvine.acc.kernels',
         'mcvine.acc.components',
         'mcvine.acc.components.sources',
         'mcvine.acc.components.optics',
@@ -26,6 +27,7 @@ setup(
         'mcvine.acc': "acc",
         'mcvine.acc.test': "acc/test",
         'mcvine.acc.geometry': "acc/geometry",
+        'mcvine.acc.kernels': "acc/kernels",
         'mcvine.acc.components': "acc/components",
         'mcvine.acc.components.sources': "acc/components/sources",
         'mcvine.acc.components.optics' : "acc/components/optics",
